Restructure project Add MoodEventViewModel
