+++ app/src/androidTest/java/com/example/mooderation/AuthenticationTest.java
@@ -10,7 +10,6 @@ import androidx.test.rule.ActivityTestRule;
 
 import com.example.mooderation.auth.ui.LoginActivity;
 import com.example.mooderation.auth.ui.SignUpActivity;
-import com.example.mooderation.viewmodel.ParticipantViewModel;
 import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
