+++ app/src/main/java/com/example/mooderation/backend/ParticipantRepository.java
@@ -44,7 +44,6 @@ public class ParticipantRepository implements Repository<Participant> {
                 .continueWithTask(task -> participantPath(participant).delete());
     }
 
-    // TODO this has a different interface than the other repos
     @Override
     public ListenerRegistration addListener(Listener<Participant> listener) {
         return participantsPath().addSnapshotListener((queryDocumentSnapshots, e) -> {
