+++ app/src/main/java/com/example/mooderation/viewmodel/MoodEventViewModel.java
@@ -1,0 +1,5 @@
+package com.example.mooderation.viewmodel;
+
+public class MoodEventViewModel {
+    // TODO
+}
