+++ app/src/main/res/layout/fragment_find_participant.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FindParticipantFragment">
+    tools:context=".fragment.FindParticipantFragment">
 
     <ListView
         android:id="@+id/search_results"
